src/VumbaSoft.AdventureWorks.Controllers/Production/ProductProductPhotos/ProductProductPhotosController.cs
using Microsoft.AspNetCore.Mvc;
using VumbaSoft.AdventureWorks.Objects;
using VumbaSoft.AdventureWorks.Services;
using VumbaSoft.AdventureWorks.Validators;
using System;

namespace VumbaSoft.AdventureWorks.Controllers.Production
{
    [Area("Production")]
    public class ProductProductPhotosController : ValidatedController<IProductProductPhotoValidator, IProductProductPhotoService>
    {
        public ProductProductPhotosController(IProductProductPhotoValidator validator, IProductProductPhotoService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create(ProductProductPhotoView photo)
        {
            if (!Validator.CanCreate(photo))
                return View(photo);

            Service.Create(photo);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(Int32 id)
        {
            return NotEmptyView(Service.Get<ProductProductPhotoView>(id));
        }

        [HttpGet]
        public ActionResult Edit(Int32 id)
        {
            return NotEmptyView(Service.Get<ProductProductPhotoView>(id));
        }

        [HttpPost]
        public ActionResult Edit(ProductProductPhotoView photo)
        {
            if (!Validator.CanEdit(photo))
                return View(photo);

            Service.Edit(photo);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(Int32 id)
        {
            return NotEmptyView(Service.Get<ProductProductPhotoView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(Int32 id)
        {
            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
